mysql connection established

+++ src/main/java/com/home/restfulwebservices/todo/TodoHardCodedService.java
@@ -10,7 +10,7 @@ public class TodoHardCodedService {
 
     private static List<Todo> todos = new ArrayList<Todo>();
 
-    private static int idCounter = 0;
+    private static long idCounter = 0;
 
     static {
         todos.add(new Todo(++idCounter,
@@ -42,7 +42,7 @@ public class TodoHardCodedService {
         return todo;
     }
 
-    public Todo deleteById(long id) {
+    public Todo deleteById(Long id) {
         Todo todo = findById(id);
         if(todo == null) return null;
 
@@ -52,7 +52,7 @@ public class TodoHardCodedService {
         return null;
     }
 
-    public Todo findById(long id) {
+    public Todo findById(Long id) {
         for(Todo todo:todos) {
             if(todo.getId() == id) {
                 return todo;
